requirements.txt
altgraph==0.17.4
chardet==5.2.0
defusedxml==0.7.1
fonttools==4.55.3
fpdf2==2.8.2
macholib==1.16.3
packaging==24.2
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
reportlab==4.2.5
setuptools==75.6.0
Unidecode==1.3.8
